src/test/java/za/co/wethinkcode/toyrobot/RobotTest.java
package za.co.wethinkcode.toyrobot;


import org.junit.jupiter.api.Test;
import za.co.wethinkcode.toyrobot.maze.SimpleMaze;
import za.co.wethinkcode.toyrobot.world.TextWorld;

import static org.junit.jupiter.api.Assertions.*;

class RobotTest {

    @Test
    void robotName() {
        Robot robot = new Robot("CrashTestDummy");
        assertEquals("CrashTestDummy", robot.getName());
    }


    @Test
    void initialPosition() {
        Robot robot = new Robot("CrashTestDummy");
        assertEquals(Robot.CENTRE, robot.getPosition());
        assertEquals(Direction.UP, robot.getCurrentDirection());
    }

    @Test
    void dump() {
        Robot robot = new Robot("CrashTestDummy");
        robot.setWorld(new TextWorld(new SimpleMaze()));
        assertEquals("[0,0] CrashTestDummy> Ready", robot.toString());
    }

    @Test
    void shutdown() {
        Robot robot = new Robot("CrashTestDummy");
        ShutdownCommand command = new ShutdownCommand();
        assertFalse(robot.handleCommand(command));
    }


    @Test
    void help() {
        Robot robot = new Robot("CrashTestDummy");
        Command command = new HelpCommand();
        assertTrue(robot.handleCommand(command));
        assertEquals("I can understand these commands:\n" +
                "OFF  - Shut down robot\n" +
                "HELP - provide information about commands\n" +
                "FORWARD - move forward by specified number of steps, e.g. 'FORWARD 10'\n" +
                "BACK - move back by specified number of steps, e.g. 'BACK 5'\n" +
                "LEFT - turns the robot to the left\n" +
                "RIGHT - turns the robot to the right\n" +
                "SPRINT - moves the robot forward quickly by specified number of steps, e.g. 'SPRINT 10'\n" +
                "REPLAY - repeats all previous movement commands", robot.getStatus());
    }
}